UkelonnServlet uses rendersnake to create the HTML.
The UkelonnServiceIntegrationTest is failing in this commit, because Pax
Exam is unable to start the bundles no.priv.bang.ukelonn_1.0.0.SNAPSHOT,
org.rendersnake_1.8.0, and org.ops4j.pax.url.war_2.4.7.

The reason the no.priv.bang.ukelonn bundle can't be resolved because the
org.rendersnake bundle isn't resolved.  But why the org.rendersnake
bundle isn't resolved is a mystery, and why the org.ops4j.pax.url.war
bundle is affected is an even bigger mystery.