Renamed the feature from "ukelonn.karaf" to "ukelonn".
This required a little refactoring, because the name of the project will
also be the name of the feature, and "ukelonn" was already taken.

Also it will be necessary to split the derby and postgresql database
plugins as separate features, since if they are loaded at the same time
it's random what database that actually will be used.