Made all features autoload the OSGi service feature.
The OSGi services are being made available as the feature ukelonn-api,
and all of the other features have this feature as a dependency.

The resulting ukelonn-all feature file is quite cluttered, because it
contains four identical copies of ukelonn-api.  However it seems to
work, so I will leave it this way.