Lazy DB fetch of jobs and payments.
Instead of fetching a list of transactions early in the Account
lifetime, and then filtering this list to create a list of jobs and a
list of payments, don't fetch either the jobs or the payments from the
database until they are needed.

And let the database do the filtering.