In the admin UI made it possible to modify an existing job type.
Inline editing in tables needed a lot of work to do what I wanted to do
so I went for a simple solution instead: populate regular forms from the
selected item in the table.

I tried using a BeanItem instead of individual ObjectProperty values,
but it wasn't possible to replace the bean of a bean item, so that
approach wasn't feasible.