Disable TouchKit offline mode.
The offline mode caused the UI to be stuck a hanging "Loading UI ..."
message when the server crashed.

However, switching off offline mode wasn't enough to fix the problem:
the loading ui message was instead replaced with an internally spinning
arrow.  This in turn is fixed in Vaadin 7.7 and newer: it is replaced
with a reaload dialog.