Try fixing the "have to click login twice problem".
The problem was that ObjectProperty<String> password contained an
empty string the first time the buttonClick() method was called.

The attempted fixes was:
 - Made the objectproperties fields instead of method local variables
 - Call password.commit() the buttonClick() method before using the value

None of these changes seemed to have any effect, but checking them in
anyway.

Note: this could be a browser specific issue. I've seen it mostly on
chrome on android, when the browser fills in remembered username and
password.