Fix test UkelonnUI test only failing with maven on debian.
The reason for the test failure was "test intereference", or lack of.
With maven on windows and in eclipse on both windows and debian, the
fake OSGi services were set up by a different test.

The fix was to make sure they are set up for the UkelonnUI tests as
well.