Make sure the job list doesn't show up empty when PostgreSQL is used as the ukelonn database.
The query didn't have a transaction_time column in the results on
PostgreSQL, but for some reason the column is present in the results
on derby.

With the transaction_time column explicitly listed, derby list two
transaction_time columns in the results, but things still work with
the column present.