Export package from OSGi bundle to avoid runtime compilation error in VaadinServlet.
The compilation error came because the compiler couldn't find the
static method CommonStringMethods.nullSafeEquals().

Side note! This class, together with the rest of the
no.priv.bang.ukelonn.impl package, can now be used in other OSGi
bundles, and the name of the package (as well as the previous lack of
package export), indicates that that was never the intention.