Exclude false positives from Sonar duplicate detection.
The AdminView/AdminFallbackView, and the
UserView/UserFallbackView are the same user interfaces with
different GUI toolkits.  They mostly contain setup code and
there is no convenient, easily maintainable way of removing
the duplications.

The PGUkelonnDatabaseProvider and UkelonnDatabaseProvider
are plugins for different RDBMS'es (PostgreSQL and Derby),
residing in separate OSGi bundles with no convenient,
maintainable way of sharing code.

Common for the two sets of "duplicates" are that they won't
run at the same time.  It's either or.