Accept only 'application/json' when handling the recent jobs and recent payments REST responses
If the page is reloaded when Shiro is in a not logged in
state (e.g. if the session has expired), what's returned from the REST
endpoints are HTML pages and jobs becomes a string instead of an
array, and rendering of the recent jobs and recent payment lists
fails.

If the response is something other than 'application/json', what's
returned is an empty array, which is the same as the initial value for
the jobs and payments in the redux state.