Avoid SQLException on /api/account/username when reloading /ukelonn/user while logged in
The SQLException caused the REST endpoint to return an Account with
a name suggesting the not logged in state, and the correct username.

The text the SQLException was: ResultSet not open. Operation 'next' not permitted. Verify that autocommit is off.

This is a Derby error, that normally is caused by the ResultSet being
closed before next() is attempted.

The fix was to remove the inner try-with-resource, since the cleanup
of the resultset is taken care of by the try-with-resource of the
PreparedStatement.