Make authservice use the same memory database as ukelonn when using a derby test database
Replace the feature authservice-db-test with ukelonn-authservice-db-test
which connects to the same derby memory database as ukelonn.

The feature ukelonn-db-test is marked as a prequiste which ensures
that the ukelonn database is created first, adds the dummy users, then
has its user tables rebuilt to the authservice schema and the data
converted.

Then ukelonn-authservice-db-test is created, connects to the same
in-memory database as ukelonn and tries to set up the authservice
liquibase schema, but discovered that the schema has been set up and
skipped and then creates a DataSource that can be injected into
authservice.