Replace pax-jdbc-config for ukelonn DB for authservice with a DS component proxying the ukelonn DataSoure OSGi service
The advantage of this solution is that there is only one config to
change, instead of two, if someone wishes to use a different memory
database than derby.

The ukelonn DataSource to authservice DataSource proxy DS component
could also, in theory, be used in production.

The disadvantage is that there is an extra DS component and an extra
OSGi module to maintain (though both the maven config and Java code
has been cut down to a minimum).